moving common data used only by the server into the server package + slowly moving over to geckos.io
