+++ client/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@dimforge/rapier3d": "0.8.1",
+    "@geckos.io/client": "^2.2.1",
     "@tweenjs/tween.js": "^18.6.4",
     "@types/resize-observer-browser": "^0.1.5",
     "@uppy/core": "2.3.1",
