+++ common/src/y-websocket/index.ts
@@ -1,5 +1,3 @@
-export * from "./callback.js";
-
 export * from "./utils.js";
 
 export * from "./y-websocket.js";
