+++ server/src/server_ws.ts
@@ -1,23 +1,11 @@
-import WebSocket from "ws";
-import { createServer } from "http";
-import { setupWSConnection } from "relm-common";
+import GeckoServer from "@geckos.io/server";
+import { encoding } from "lib0";
+import * as syncProtocol from "y-protocols/sync";
+import * as awarenessProtocol from "y-protocols/awareness";
 
-import { app } from "./server_http.js";
 import { Participant, Permission, Doc } from "./db/index.js";
 import { ydocStats } from "./getYDoc.js";
-import { hasPermission } from "./utils/hasPermission.js";
-
-export const server = createServer();
-
-let wss = new WebSocket.Server({ noServer: true });
-
-wss.on("connection", (conn, req) => {
-  setupWSConnection(conn, req, {
-    callbackHandler: ydocStats,
-  });
-});
-
-server.on("request", app);
+import { hasPermission, getYDoc } from "./utils/index.js";
 
 function getRelmDocFromRequest(req) {
   return req.url.slice(1).split("?")[0];
@@ -28,31 +16,40 @@ function getUrlParams(requestUrl) {
   return new URLSearchParams(queryString);
 }
 
-server.on("upgrade", async (req, socket, head) => {
-  const docId = getRelmDocFromRequest(req);
-  const params = getUrlParams(req.url);
+const messageSync = 0;
+const messageAwareness = 1;
 
-  const participantId = params.get("participant-id");
-  const participantSig = params.get("participant-sig");
-  let pubkeyX = params.get("pubkey-x");
-  let pubkeyY = params.get("pubkey-y");
-  console.log("participant connected:", participantId);
+export const geckoServer = GeckoServer({
+  cors: { allowAuthorization: true, origin: "*" }, // Required since client is on separate domain
 
-  let verifiedPubKey;
-  try {
-    verifiedPubKey = await Participant.findOrCreateVerifiedPubKey({
-      participantId,
-      sig: participantSig,
-      x: pubkeyX,
-      y: pubkeyY,
-    });
-  } catch (err) {
-    console.warn("can't upgrade", err);
-    return;
-  }
+  authorization: async (_, req) => {
+    const docId = getRelmDocFromRequest(req);
+    const params = getUrlParams(req.url);
+  
+    const participantId = params.get("participant-id");
+    const participantSig = params.get("participant-sig");
+    let pubkeyX = params.get("pubkey-x");
+    let pubkeyY = params.get("pubkey-y");
+    console.log("participant connected:", participantId);
+  
+    let verifiedPubKey;
+    try {
+      verifiedPubKey = await Participant.findOrCreateVerifiedPubKey({
+        participantId,
+        sig: participantSig,
+        x: pubkeyX,
+        y: pubkeyY,
+      });
+    } catch (err) {
+      console.warn("can't upgrade", err);
+      return;
+    }
+  
+    // Check that we are authenticated first
+    if (!verifiedPubKey) {
+      return false;
+    }
 
-  // Check that we are authenticated first
-  if (verifiedPubKey) {
     // Get relm from docId
     const doc = await Doc.getDoc({ docId });
 
@@ -61,28 +58,65 @@ server.on("upgrade", async (req, socket, head) => {
       relmIds: [doc.relmId],
     });
 
+    // Do we have permission to access this relm?
     const permitted = hasPermission("access", permissions, doc.relmId);
 
-    if (permitted) {
-      if (doc === null) {
-        console.log(
-          `Participant '${participantId}' sought to sync doc '${docId}' but was rejected because it doesn't exist`
-        );
-        socket.write("HTTP/1.1 404 Not Found\r\n\r\n");
-        socket.destroy();
-      } else {
-        wss.handleUpgrade(req, socket as any, head, (conn) => {
-          wss.emit("connection", conn, req);
-        });
-      }
-    } else {
+    if (!permitted) {
       console.log(
         `Participant '${participantId}' sought to enter '${docId}' but was rejected because unauthorized`,
         params,
         permissions
       );
-      socket.write("HTTP/1.1 401 Unauthorized\r\n\r\n");
-      socket.destroy();
+      return 401;
     }
+
+    // Does the relm doc exist?
+    if (doc === null) {
+      console.log(
+        `Participant '${participantId}' sought to sync doc '${docId}' but was rejected because it doesn't exist`
+      );
+      return 404;
+    }
+
+    // Good to go!
+    return {
+      docId
+    };
+  }
+});
+
+geckoServer.onConnection(async channel => {
+  const doc = await getYDoc(channel.userData.docId, { callbackHandler: ydocStats });
+  doc.conns.set(channel, new Set());
+
+  channel.onRaw(event => {
+    console.log(event);
+    // messageListener(channel, doc, new Uint8Array(message))
+  });
+
+  channel.onDisconnect(() => {
+    console.log("DISCONNECT DETECTED ON CHANNEL. SHOULD CLOSE CONNECTION");
+  });
+
+  if (doc.whenSynced) {
+    await doc.whenSynced;
+  }
+
+  const encoder = encoding.createEncoder();
+  encoding.writeVarUint(encoder, messageSync);
+  syncProtocol.writeSyncStep1(encoder, doc);
+  channel.raw.emit(encoding.toUint8Array(encoder));
+  const awarenessStates = doc.awareness.getStates();
+  if (awarenessStates.size > 0) {
+    const encoder = encoding.createEncoder();
+    encoding.writeVarUint(encoder, messageAwareness);
+    encoding.writeVarUint8Array(
+      encoder,
+      awarenessProtocol.encodeAwarenessUpdate(
+        doc.awareness,
+        Array.from(awarenessStates.keys())
+      )
+    );
+    channel.raw.emit(encoding.toUint8Array(encoder));
   }
 });
