+++ server/src/ws/WSSharedDoc.ts
@@ -1,0 +1,121 @@
+import * as Y from "yjs";
+import * as syncProtocol from "y-protocols/sync";
+import * as awarenessProtocol from "y-protocols/awareness";
+import { mutex, encoding } from "lib0";
+import debounce from "lodash.debounce";
+import type { ServerChannel } from "@geckos.io/server";
+
+import { YGARBAGE_COLLECTION } from "../config.js";
+import { defaultCallbackHandler, isDefaultCallbackSet } from "./callback.js";
+
+const CALLBACK_DEBOUNCE_WAIT =
+  parseInt(process.env.CALLBACK_DEBOUNCE_WAIT) || 2000;
+const CALLBACK_DEBOUNCE_MAXWAIT =
+  parseInt(process.env.CALLBACK_DEBOUNCE_MAXWAIT) || 10000;
+
+const messageSync = 0;
+const messageAwareness = 1;
+
+/**
+ * @param {Uint8Array} update
+ * @param {any} origin
+ * @param {WSSharedDoc} doc
+ */
+ const updateHandler = (update, origin, doc) => {
+  const encoder = encoding.createEncoder();
+  encoding.writeVarUint(encoder, messageSync);
+  syncProtocol.writeUpdate(encoder, update);
+  const message = encoding.toUint8Array(encoder);
+  doc.conns.forEach((_, conn) => {
+    conn.raw.send(message);
+  });
+  // doc.conns.forEach((_, conn) => send(doc, conn, message));
+};
+
+export class WSSharedDoc extends Y.Doc {
+  name: string;
+  mux: any;
+  conns: Map<ServerChannel, Set<number>>;
+  awareness: any;
+  whenSynced: Promise<void>;
+  isSynced: boolean;
+
+  /**
+   * @param {string} name
+   */
+  constructor(
+    name,
+    { callbackHandler = isDefaultCallbackSet ? defaultCallbackHandler : null, persistence } : { callbackHandler: any, persistence? }
+  ) {
+    super({ gc: YGARBAGE_COLLECTION });
+    this.name = name;
+    this.mux = mutex.createMutex();
+    /**
+     * Maps from conn to set of controlled user ids. Delete all user ids from awareness when this conn is closed
+     * @type {Map<ServerChannel, Set<number>>}
+     */
+    this.conns = new Map();
+    /**
+     * @type {awarenessProtocol.Awareness}
+     */
+    this.awareness = new awarenessProtocol.Awareness(this);
+    this.awareness.setLocalState(null);
+    /**
+     * @type {Promise<void>|void}
+     */
+    this.whenSynced = void 0;
+    /**
+     * @type {boolean}
+     */
+    this.isSynced = false;
+    /**
+     * @param {{ added: Array<number>, updated: Array<number>, removed: Array<number> }} changes
+     * @param {Object | null} conn Origin is the connection that made the change
+     */
+    const awarenessChangeHandler = ({ added, updated, removed }, conn) => {
+      const changedClients = added.concat(updated, removed);
+      if (conn !== null) {
+        const connControlledIDs =
+          /** @type {Set<number>} */ this.conns.get(conn);
+        if (connControlledIDs !== undefined) {
+          added.forEach((clientID) => {
+            connControlledIDs.add(clientID);
+          });
+          removed.forEach((clientID) => {
+            connControlledIDs.delete(clientID);
+          });
+        }
+      }
+      // broadcast awareness update
+      const encoder = encoding.createEncoder();
+      encoding.writeVarUint(encoder, messageAwareness);
+      encoding.writeVarUint8Array(
+        encoder,
+        awarenessProtocol.encodeAwarenessUpdate(this.awareness, changedClients)
+      );
+      const buff = encoding.toUint8Array(encoder);
+      this.conns.forEach((_, c) => {
+        c.raw.emit(buff);
+        // send(this, c, buff);
+      });
+    };
+    this.awareness.on("update", awarenessChangeHandler);
+    this.on("update", updateHandler);
+
+    if (callbackHandler) {
+      this.on(
+        "update",
+        debounce(callbackHandler, CALLBACK_DEBOUNCE_WAIT, {
+          maxWait: CALLBACK_DEBOUNCE_MAXWAIT,
+        })
+      );
+    }
+
+    if (persistence !== null) {
+      console.log(persistence);
+      this.whenSynced = persistence.bindState(name, this).then(() => {
+        this.isSynced = true;
+      });
+    }
+  }
+}
