+++ server/src/ws/callback.ts
@@ -1,4 +1,5 @@
 import http from "http";
+import {  } from "../config.js";
 
 const CALLBACK_URL = process.env.CALLBACK_URL
   ? new URL(process.env.CALLBACK_URL)
